[ListPad-1] Basic folder structure refactor
