+++ app/src/main/java/vander/gabriel/listpad/presentation/theme/Color.kt
@@ -1,4 +1,4 @@
-package vander.gabriel.listpad.ui.theme
+package vander.gabriel.listpad.presentation.theme
 
 import androidx.compose.ui.graphics.Color
